requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
PyJWT==1.6.1
requests==2.18.4
Salesforce-FuelSDK==1.0.1
simple-salesforce==0.74.2
six==1.11.0
suds-jurko==0.6
teradata==15.10.0.21
unicodecsv==0.14.1
urllib3==1.22
